# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The central API Gateway for the Adeff distributed audiobook platform.",
        "title": "Adeff API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/admin/upload": {
            "post": {
                "description": "Initiates the Saga pattern. Uploads PDF to MinIO, creates DB record, and triggers the Analyzer worker.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Upload a PDF to trigger audio generation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of the book",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language (e.g., HINDI, ENGLISH, MIXED)",
                        "name": "language",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "The PDF file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the operational status of the API Gateway",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Check system health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}